move data retrival for info fields
